Modified logic in DeleteReviewsServlet.java class

+++ src/main/java/servlets/DeleteReviewsServlet.java
@@ -18,32 +18,40 @@ import java.util.List;
 public class DeleteReviewsServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        PrintWriter out = response.getWriter();
-        String hotelId = request.getParameter("hotelId");
-        System.out.println(hotelId);
-        response.setContentType("text/html");
-        response.setStatus(HttpServletResponse.SC_OK);
         HttpSession session = request.getSession();
-        String username = (String)session.getAttribute("username");
-        System.out.println(username);
-        DatabaseHandler handler = DatabaseHandler.getInstance();
-        List<Review> reviewList = handler.getReviewsForUser(hotelId,username);
-        ThymeLeafConfig thymeleafConfig = new ThymeLeafConfig();
-        ThymeLeafRenderer thymeleafRenderer = new ThymeLeafRenderer(thymeleafConfig.templateEngine());
-        thymeleafRenderer.setVariable("reviews",reviewList);
-        thymeleafRenderer.setVariable("hotelId",hotelId);
-        thymeleafRenderer.render("deleteReviews", out);
+        if (session.getAttribute("username") == null) {
+            response.sendRedirect("/login");
+        }else {
+            PrintWriter out = response.getWriter();
+            String hotelId = request.getParameter("hotelId");
+            String hotelName = request.getParameter("hotelName");
+            response.setContentType("text/html");
+            response.setStatus(HttpServletResponse.SC_OK);
+            String username = (String) session.getAttribute("username");
+            DatabaseHandler handler = DatabaseHandler.getInstance();
+            List<Review> reviewList = handler.getReviewsForUser(hotelId, username);
+            ThymeLeafConfig thymeleafConfig = new ThymeLeafConfig();
+            ThymeLeafRenderer thymeleafRenderer = new ThymeLeafRenderer(thymeleafConfig.templateEngine());
+            thymeleafRenderer.setVariable("reviews", reviewList);
+            thymeleafRenderer.setVariable("hotelId", hotelId);
+            thymeleafRenderer.setVariable("hotelName", hotelName);
+            thymeleafRenderer.render("deleteReviews", out);
+        }
     }
 
     @Override
-    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        String reviewId = req.getParameter("reviewId");
-        String hotelId = req.getParameter("hotelId");
-        System.out.println("reviewId "+ reviewId);
-        System.out.println("hotelId "+ hotelId);
-        DatabaseHandler handler = DatabaseHandler.getInstance();
-        handler.deleteReview(reviewId,hotelId);
-        resp.sendRedirect("/deleteReviews?hotelId="+hotelId);
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        String reviewId = request.getParameter("reviewId");
+        String hotelId = request.getParameter("hotelId");
+        String hotelName = request.getParameter("hotelName");
+        String action = request.getParameter("action");
+        if (action.equals("delete")) {
+            DatabaseHandler handler = DatabaseHandler.getInstance();
+            handler.deleteReview(reviewId,hotelId);
+            response.sendRedirect("/deleteReviews?hotelId="+hotelId+"&hotelName="+hotelName);
+        } else if (action.equals("edit")) {
+            response.sendRedirect("/editReviews?hotelId="+hotelId+"&reviewId="+reviewId);
+        }
 
     }
 }
